pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "google-generativeai>=0.8.3",
    "requests>=2.32.3",
]
